refactor: remove .env.example and implement Modal integration for PDF OCR settings and client
- Deleted the .env.example file as it is no longer needed.
- Added new layout.py for PDF OCR settings using Pydantic.
- Created marker_client.py to handle interactions with the Modal-hosted Marker service.
- Updated ocr_jobs.py to import the new Modal client for document conversion.
