--- .env.example
@@ -1,6 +1,0 @@
-# Marker service configuration
-MARKER_RUN_MODE=modal
-MARKER_MODAL_BASE_URL=https://YOUR-ENDPOINT.modal.run
-
-# Optional tuning
-# MARKER_MODAL_TIMEOUT_SECS=600
